+++ models/reply_markup.go
@@ -74,13 +74,13 @@ type KeyboardButtonPollType struct {
 }
 
 type CallbackGame struct {
-	UserID             int  `json:"user_id"`
-	Score              int  `json:"score"`
-	Force              bool `json:"force,omitempty"`
-	DisableEditMessage bool `json:"disable_edit_message,omitempty"`
-	ChatID             int  `json:"chat_id,omitempty"`
-	MessageID          int  `json:"message_id,omitempty"`
-	InlineMessageID    int  `json:"inline_message_id,omitempty"`
+	UserID             int64 `json:"user_id"`
+	Score              int   `json:"score"`
+	Force              bool  `json:"force,omitempty"`
+	DisableEditMessage bool  `json:"disable_edit_message,omitempty"`
+	ChatID             int   `json:"chat_id,omitempty"`
+	MessageID          int   `json:"message_id,omitempty"`
+	InlineMessageID    int   `json:"inline_message_id,omitempty"`
 }
 
 // ReplyKeyboardRemove https://core.telegram.org/bots/api#replykeyboardremove
